+++ Examples/Io/HepMC3/src/HepMC3Event.cpp
@@ -6,10 +6,10 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/Plugins/HepMC3/HepMC3Event.hpp"
+#include "ActsExamples/Io/HepMC3/HepMC3Event.hpp"
 
-#include "ActsExamples/Plugins/HepMC3/HepMC3Particle.hpp"
-#include "ActsExamples/Plugins/HepMC3/HepMC3Vertex.hpp"
+#include "ActsExamples/Io/HepMC3/HepMC3Particle.hpp"
+#include "ActsExamples/Io/HepMC3/HepMC3Vertex.hpp"
 
 ///
 /// Setter
