+++ Examples/Io/HepMC3/src/HepMC3Reader.cpp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "ActsExamples/Plugins/HepMC3/HepMC3Reader.hpp"
+#include "ActsExamples/Io/HepMC3/HepMC3Reader.hpp"
 
 bool ActsExamples::HepMC3ReaderAscii::readEvent(
     HepMC3::ReaderAscii& reader, std::shared_ptr<HepMC3::GenEvent> event) {
